refactor: 사용하지않는 레거시 BeanFactory 클래스 삭제
